put tests in own folder
