+++ testing/__init__.py
@@ -1,0 +1,6 @@
+from .tests import test_nucleus_sampling, test_overfitting_study
+
+__all__ = [
+    "test_nucleus_sampling",
+    "test_overfitting_study"
+]
